refactor(webui): organize markdown editor files into components/markdown-editor folder
