+++ packages/webui/components/markdown-editor/index.ts
@@ -1,0 +1,2 @@
+export { MarkdownEditor, type MarkdownEditorProps } from './markdown-editor'
+export { default } from './markdown-editor'
